Add module for the comments
