+++ src/modules/popup.js
@@ -1,3 +1,5 @@
+import addComment from "./Comments.js";
+
 const main = document.querySelector('header');
 
 const urlAPI = 'https://pokeapi.co/api/v2/pokemon';
@@ -34,6 +36,19 @@ const createPopup = (pokeInfo) => {
   close.addEventListener('click', () => {
     main.removeChild(cards);
   });
+
+  const commentBtn = document.querySelector('.commentButton');
+
+commentBtn.addEventListener('click', (event) => {
+event.preventDefault();
+const { index } = event.target.dataset;
+const takeUser = document.querySelector('.username').value;
+const takeComment = document.querySelector('.comment').value;
+
+addComment(index, takeUser, takeComment);
+document.querySelector('.comment').value = "";
+document.querySelector('.username').value = "";
+});
 };
 
 // fetch API
@@ -66,51 +81,5 @@ const getPokeInfo = () => {
     }
   });
 };
-
-
-//Working on Involvement API
-const idComments = 'BZDVsIE83KOtqyh07flb'
-
-//post comment
-const addComment = async (id, username, comment) => {
-  try {
-    const response = await fetch("https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/zVvcayqMcKSodjVAeGk6/comments", {
-      method: 'POST',
-      headers: {
-        'Content-type': 'application/json; charset=UTF-8',
-      },
-      body: JSON.stringify({
-        "item_id": id,
-        "username": username,
-        "comment": comment
-      })      
-    })
-    
-    const data = await response.text();
-    console.log(data);
-    return data;
-  } catch (error) {
-    throw new Error('Game not found!');
-  }
-};
-
-
-// Get users comments
-const userComments = async (id) => {
-  try {
-    const response = await fetch(`https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/zVvcayqMcKSodjVAeGk6/comments?item_id=${id}`);
-    const data = await response.json();
-    console.log(data)
-    return data;
-  } catch (error) {
-    throw new Error('Not Found!!!.');
-  }
-};
-
-userComments('item2')
-
-
-
-
 
 export default getPokeInfo;
